requirements.txt
fastapi
uvicorn[standard]
opencv-python==4.9.0.80
mediapipe==0.10.11
numpy
fpdf2
faster-whisper
soundfile
python-multipart
